requirements.txt
fastmcp>=3.0.0
chromadb>=1.0.0
sentence-transformers>=5.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
